+++ docs/CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 - Add a pseudo Python process to serve as a template
 - Add ``validate_urls`` method to ``BaseConnector`` class.
 - Validate connector settings on registraton
+- Add ``transfer_data`` method to ``StorageLocation`` class
 
 Fixed
 -----